requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
flask
pypdf
pymupdf
python-dotenv
langchain-pinecone
langchain-openai
langchain==0.1.14
-e .